move grid renderer into a separate entity
